requirements.txt
fastmcp>=1.0.0
chromadb>=0.4.0
GitPython>=3.1.44
python-dotenv>=1.0.0
anthropic>=0.5.0
openai>=1.0.0